Refacto contexts: custom hooks placed directly inside their related Context file
