--- frontend/src/hooks/useThemeContext.jsx
@@ -1,16 +1,0 @@
-import { useContext } from "react";
-import { ThemeContext } from "../contexts/ThemeContext";
-
-const userThemeContext = () => {
-  // get the context
-  const context = useContext(ThemeContext);
-
-  // if `undefined`, throw an error
-  if (context === undefined) {
-    throw new Error("userThemeContext was used outside of its Provider!");
-  }
-
-  return context;
-};
-
-export default userThemeContext;
