+++ frontend/src/main.jsx
@@ -3,6 +3,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter as Router } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
+
 import { ThemeContextProvider } from "./contexts/ThemeContext";
 import { AuthContextProvider } from "./contexts/AuthContext";
